% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bic.glmnet}
\alias{bic.glmnet}
\title{BIC for glmnet}
\usage{
bic.glmnet(X, y, nlambda = 100)
}
\arguments{
\item{X}{design matrix}

\item{y}{response vector}

\item{nlambda}{number of lambda values}
}
\value{
The estimated coefficient vector selected by BIC
}
\description{
Compute BIC for glmnet and return the estimated coefficient
}
