% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calN_j}
\alias{calN_j}
\title{Calculate indices for node j}
\usage{
calN_j(n, j)
}
\arguments{
\item{n}{number of samples per node}

\item{j}{node index}
}
\value{
A vector of indices assigned to node j
}
\description{
Calculate the indices of samples assigned to node j
}
