% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{computeError}
\alias{computeError}
\title{Compute error between two vectors or two matrices}
\usage{
computeError(x, y, type = "F")
}
\arguments{
\item{x}{first vector or matrix}

\item{y}{second vector or matrix}

\item{type}{norm type, default is "F" (Frobenius norm)}
}
\value{
The computed error
}
\description{
Compute the error between two vectors or two matrices using
specified norm.
}
