% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{computeF1}
\alias{computeF1}
\title{Compute F1 score of two set}
\usage{
computeF1(esupp, supp)
}
\arguments{
\item{esupp}{the estimated support}

\item{supp}{the true support}
}
\value{
A list consists of
\item{f1}{the computed F1 score}
\item{precision}{the precision}
\item{recall}{the recall}
\item{fp}{false positve}
\item{fn}{false negative}
}
\description{
Compute \eqn{F_1} score of two set. Note that \eqn{F_1} score is
symmetric.
}
\references{
\url{https://en.wikipedia.org/wiki/F-score}
}
