% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createdir}
\alias{createdir}
\title{Create directory if not exist}
\usage{
createdir(path, recursive = TRUE)
}
\arguments{
\item{path}{directory path}

\item{recursive}{logical, whether to create directories recursively}
}
\value{
TRUE if directory is created, FALSE otherwise
}
\description{
Create directory if not exist
}
