% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fixRNGStream}
\alias{fixRNGStream}
\title{Fix RNG stream for parallel computing}
\usage{
fixRNGStream(s, batch)
}
\arguments{
\item{s}{initial RNG stream}

\item{batch}{number of streams to advance}
}
\value{
The fixed RNG stream after certain number of streams.
}
\description{
Given an initial RNG stream, fix the RNG stream after certain
number of streams.
}
