% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{genData}
\alias{genData}
\title{Generate synthetic 1-bit data}
\usage{
genData(
  N = 2000,
  m = 10,
  p = 100,
  s = 5,
  pc = 0.3,
  p_flip = 0.2,
  rho = 0.1,
  sigma2 = 1,
  type = c("Cauchy", "Normal", "T2"),
  sigma2_set = c(1, 3),
  rho_set = c(0.1, 0.3),
  p_flip_set = c(0.05, 0.1),
  case = 1
)
}
\arguments{
\item{N}{total sample size}

\item{m}{number of nodes}

\item{p}{number of features}

\item{s}{sparsity level of the true signal}

\item{pc}{connection probability for the network graph}

\item{p_flip}{flipping probability for the 1-bit measurements}

\item{rho}{correlation parameter for the design matrix}

\item{sigma2}{variance parameter for the design matrix}

\item{type}{error distribution type, can be "Cauchy", "Normal" or "T2"}

\item{sigma2_set}{set of variance parameters for heterogeneous data}

\item{rho_set}{set of correlation parameters for heterogeneous data}

\item{p_flip_set}{set of flipping probabilities for heterogeneous data}

\item{case}{data generation case, can be 0, 1, 2 or 3}
}
\value{
A list consists of
\item{X}{design matrix}
\item{y}{response vector}
\item{betaT}{true signal vector}
\item{graph}{network graph}
}
\description{
Generate synthetic 1-bit data for decentralized 1-bit
compressed sensing.
}
