% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kernel}
\alias{kernel}
\title{Kernel fitting}
\usage{
kernel(E, h)
}
\arguments{
\item{E}{A n-by-K residual matrix, where n is sample size and K is the
quantile levels}

\item{h}{A n-by-K bandwidth}
}
\value{
A K-by-1 density estimate vector.
}
\description{
Conduct the Nadaraya-Watson kernel regression to estimate the density of the
errs.
}
\examples{
set.seed(1)
library(pracma)
p <- 600
n <- 100
case <- "cauchy"
betaT <- rep(0, p)
betaT[1] <- 3
betaT[2] <- 1.5
betaT[5] <- 2
s <- length(which(abs(betaT)>0))
rho <- 0.5
data <- gen_data(n, p, betaT, case, rho)
x <- data$x
y <- data$y
data_val <- gen_data(n, p, betaT, case, rho)
x_val <- data_val$x
y_val <- data_val$y
K <- 19
tau <- 0.05 * (1:K)
m <- qraenet(x = x, y = y, tau = 0.5, standardize = FALSE,
intercept = FALSE, lambda2 = 0, sigma = 1, method = "padmm")
nlambda <- length(m$lambda)
validation_err <- rep(NA, nlambda)
for(ivalidation_err in 1:nlambda) {
  validation_err[ivalidation_err] <-
    sum((y_val - x_val\%*\%m$beta[,ivalidation_err])^2)
}
beta_init <-  matrix(as.numeric(m$beta[,which.min(validation_err)]))
alpha_init <- quantile(y - x\%*\%beta_init, tau)
E <- repmat(y - x\%*\%beta_init, 1, K) - repmat(alpha_init, n, 1);
h <- rep(sqrt(s*log(n)/n), K)
fh <- kernel(E, h)
}
