% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mod}
\alias{mod}
\title{Mode of a given vector}
\usage{
mod(v)
}
\arguments{
\item{v}{A vector of real numbers}
}
\value{
the mode of the vector
}
\description{
Find the mode of a given numerical vector.
}
\examples{
v <- c(rep(1,10),rep(2,9))
mod(v)
}
